Moved insurance accident totals calculation to useInsuranceAccidentTotals hook
